test: move user task variable update tests to TaskListenerVariablesTest
Migrated tests from `UserTaskVariableUpdateTest` to `TaskListenerVariablesTest`,
aligning with the recent refactoring of task listener tests.

Deleted `UserTaskVariableUpdateTest` since its functionality is now covered
by the new structure.

Added `getUserTaskElementInstanceKey` utility method to `TaskListenerTestHelper`
to facilitate retrieving the user task element instance key in tests.
